One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/planning/MerchSetup/addmerch/addmerch.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { AddmerchComponent } from './addmerch.component';

describe('Component: Addmerch', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [AddmerchComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([AddmerchComponent],
      (component: AddmerchComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(AddmerchComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(AddmerchComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-addmerch></app-addmerch>
  `,
  directives: [AddmerchComponent]
})
class AddmerchComponentTestController {
}

